♻️ Move Level to opts and add temporary alias to sqlslog package
